Refactoring of Server with new class RoomService
RoomService takes care of adding and deleting clients, as well as
assignin master and controllers.
[pass] connect master
[pass] disconnect
[pass] connect controller to master
[pass] disconnect controller to master
[pass] send message from controller to master
[pass] send message from master to controller

Sending messages is in a sort of "private space", i.e. If room 0002
sends message to player1, i.e. the first client connected to room 0002.
The room 0001 will not see any of the conversation.

No function implemented to communicate between room. However, it is
possible.